Split search stores into own files
