+++ src/modules/data/search/organizer/selectors.js
@@ -1,0 +1,7 @@
+export function fetchDetails( state, organizers ) {
+	return state;
+}
+
+export function search( state ) {
+	return state.search;
+}
